Added bowtie module
