+++ multiqc/base_module.py
@@ -18,11 +18,13 @@ letters = 'abcdefghijklmnopqrstuvwxyz'
 
 class BaseMultiqcModule(object):
 
-    def __init__(self, name='base', anchor='bismark', href='', info=''):
+    def __init__(self, name='base', anchor='bismark', target='',href='', info=''):
         self.name = name
         self.anchor = anchor
+        if not target:
+            target = self.name
         self.intro = '<p><a href="{0}" target="_blank">{1}</a> {2}</p>'.format(
-            href, self.name, info
+            href, target, info
         )
 
     def find_log_files(self, fn_match=None, contents_match=None, filehandles=False):
